fixing issue in ch.julien.query.core.TraversableImpl#take - livelock with infinite iterable

+++ common/src/main/java/ch/julien/common/monad/Indexed.java
@@ -1,16 +1,16 @@
 package ch.julien.common.monad;
 
 public class Indexed<T> {
-	private final int index;
+	private final long index;
 	private final T value;
 
 
-	public Indexed(int index, T value) {
+	public Indexed(long index, T value) {
 		this.index = index;
 		this.value = value;
 	}
 
-	public int getIndex() {
+	public long getIndex() {
 		return this.index;
 	}
